# Scheduler-integrated/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexusDesk — NITK ECE Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /semesters:
    get:
      operationId: listSemesters
      tags: [semesters]
      summary: List all semesters
      responses:
        "200":
          description: List of semesters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Semester"
    post:
      operationId: createSemester
      tags: [semesters]
      summary: Create a semester
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemesterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Semester"
  /semesters/active:
    get:
      operationId: getActiveSemester
      tags: [semesters]
      summary: Get the active semester
      responses:
        "200":
          description: Active semester
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Semester"
        "404":
          description: No active semester
  /semesters/{semesterId}:
    get:
      operationId: getSemester
      tags: [semesters]
      summary: Get a semester by ID
      parameters:
        - name: semesterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Semester
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Semester"
        "404":
          description: Not found
    patch:
      operationId: updateSemester
      tags: [semesters]
      summary: Update a semester
      parameters:
        - name: semesterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SemesterUpdate"
      responses:
        "200":
          description: Updated semester
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Semester"

  # ─── COURSES ──────────────────────────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events for a date range
      parameters:
        - name: date
          in: query
          schema:
            type: string
          description: ISO date string for the target day
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventWithAttendance"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event(s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEventResult"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithAttendance"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithAttendance"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /events/{eventId}/cancel:
    post:
      operationId: cancelEvent
      tags: [events]
      summary: Cancel an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelEventInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithAttendance"

  # ─── ATTENDANCE ───────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks, optionally filtered
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── HARDWARE PROJECTS ────────────────────────────────────────────────
...
components:
  schemas:
...
    Semester:
      type: object
      required: [id, name, startDate, endDate, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    SemesterInput:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
    SemesterUpdate:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean

    # ── Course ──
...
    EventWithAttendance:
      type: object
      required: [id, title, type, startTime, endTime, isCancelled, isRecurring, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: ["string", "null"]
        isCancelled:
          type: boolean
        cancellationNote:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        recurringGroupId:
          type: ["string", "null"]
        courseId:
          type: ["string", "null"]
        courseShortName:
          type: ["string", "null"]
        courseColor:
          type: ["string", "null"]
        attendanceStatus:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    EventInput:
      type: object
      required: [title, type, startTime, endTime]
      properties:
        title:
          type: string
        type:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        courseId:
          type: string
        isRecurring:
          type: boolean
        recurrenceDays:
          type: array
          items:
            type: integer
        recurrenceUntil:
          type: string
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        startTime:
          type: string
        endTime:
          type: string
    CancelEventInput:
      type: object
      properties:
        cancellationNote:
          type: string
    CreateEventResult:
      type: object
      required: [created]
      properties:
        created:
          type: integer
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventWithAttendance"

    # ── Attendance ──
...
    Task:
      type: object
      required: [id, title, status, category, priority, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
        category:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        linkedCourseId:
          type: ["string", "null"]
        linkedProjectId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [title, category]
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        category:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        linkedCourseId:
          type: string
        linkedProjectId:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        status:
          type: string
        category:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string

    # ── Project ──
...