# Scheduler-integrated/lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NexusDesk — NITK ECE Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  title: string;
  type: string;
  startTime: string;
  endTime: string;
  location?: string;
  courseId?: string;
  isRecurring?: boolean;
  recurrenceDays?: number[];
  recurrenceUntil?: string;
}